Regression: Remove lazy function body parsing
Lazy function body parsing requires, at least for the old implementation, to store the body tokens in the AST Decl node. However, AST nodes are allocated by a bump allocator and are never freed nor their destructor are called. Thus using a std::vector is not viable, so a solution must be found. However, this is not an actual priority.